KA10: IMX device stub.
This is a stub for device 574, which is a 12-bit A/D converter with
128 channels.  There was only one, attached to the MIT AI lab PDP-10.
It's documented in ITS hardware memo #4.

This implementation deals with CONO and CONI bits, and accepts a
channel number written with DATAO.  However, the data returned by
DATAI is always 2048.  The programmed data rate is ignored and data is
always available immediately.  It will not generate interrupts.