KA10: Tymnet interface.
A terminal multiplexing device using the TENEX Tymbase shared memory
interface.

For now, this only implements a bare minimum of Tymbase messages:

From the host to the Tymbase, messages 01 (host open) or 21 (set host
number) activates input from terminal lines.  From the Tymbase to the
host, message 06 (login) is sent when a terminal line is connected and
message 13 (zap) is sent when a line is disconnected.  Between both,
data messages are exchange for character data.